requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==23.1.0
bleach==6.2.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
conda==23.1.0
conda-content-trust @ file:///tmp/abs_5952f1c8-355c-4855-ad2e-538535021ba5h26t22e5/croots/recipe/conda-content-trust_1658126371814/work
conda-package-handling @ file:///croot/conda-package-handling_1672865015732/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508151586/work
cryptography @ file:///croot/cryptography_1673298753778/work
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flake8==6.0.0
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna @ file:///croot/idna_1666125576474/work
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lightning-utilities==0.11.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.11.0
pexpect==4.9.0
pickleshare==0.7.5
pipenv==2023.2.18
pipreqs==0.5.0
platformdirs==3.0.0
pluggy @ file:///tmp/build/80754af9/pluggy_1648024709248/work
prompt_toolkit==3.0.48
propcache==0.2.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.10.0
pycosat @ file:///croot/pycosat_1666805502580/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==3.0.1
Pygments==2.18.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
rpds-py==0.21.0
ruamel.yaml @ file:///croot/ruamel.yaml_1666304550667/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302247304/work
safetensors==0.4.5
semantic-version==2.10.0
setuptools-rust==1.10.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
torch==1.13.1
torchmetrics==1.5.2
tornado==6.4.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
traitlets==5.14.3
transformers==4.46.3
typing_extensions==4.12.2
urllib3 @ file:///croot/urllib3_1673575502006/work
virtualenv==20.20.0
virtualenv-clone==0.5.7
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.18.0
zstandard @ file:///opt/conda/conda-bld/zstandard_1663827383994/work
